compute/python/requirements.txt
python-dotenv
python-openstackclient